notification-service/requirements.txt
flask
pymysql
mysql-connector-python
python-dotenv